Incorporated Duke's Comments
