+++ pkg/sloop/server/server.go
@@ -132,7 +132,7 @@ func RealMain() error {
 			BadgerVLogGCFreq:   conf.BadgerVLogGCFreq,
 			DeletionBatchSize:  conf.DeletionBatchSize,
 			GCThreshold:        conf.ThresholdForGC,
-			EnableDeletePrefix: conf.EnableDeleteKeys,
+			EnableDeleteKeys:   conf.EnableDeleteKeys,
 		}
 		storemgr = storemanager.NewStoreManager(tables, storeCfg, fs)
 		storemgr.Start()
